pyproject.toml
[project]
name = "ndc-sdk-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Daniel Harvey", email = "danieljamesharvey@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.7.0,<3.0.0",
  "typing_extensions>=4.11.0,<5.0.0",
  "fastapi>=0.110.2",
  "uvicorn>=0.29.0",
  "opentelemetry-api>=1.20.0,<2.0.0",
  "opentelemetry-exporter-otlp>=1.20.0,<2.0.0",
  "opentelemetry-instrumentation-fastapi>=0.44b0",
  "opentelemetry-instrumentation-logging>=0.44b0",
  "opentelemetry-instrumentation-requests>=0.44b0",
  "opentelemetry-propagator-b3>=1.20.0,<2.0.0"
]

[project.scripts]
ndc-sdk-python = "ndc_sdk_python:main"

[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"
